+++ apps/web/src/app/(signing)/sign/[token]/complete/page.tsx
@@ -9,8 +9,8 @@ import { getFieldsForToken } from '@documenso/lib/server-only/field/get-fields-f
 import { getRecipientByToken } from '@documenso/lib/server-only/recipient/get-recipient-by-token';
 import { DocumentStatus, FieldType } from '@documenso/prisma/client';
 import { DocumentDownloadButton } from '@documenso/ui/components/document/document-download-button';
+import { SigningCard } from '@documenso/ui/components/signing-card';
 import { Button } from '@documenso/ui/primitives/button';
-import { SigningCard } from '@documenso/ui/primitives/signing-card';
 
 import signingCelebration from '~/assets/signing-celebration.png';
